Fixed benchmark queries and module layout
