+++ juniper_benchmarks/Cargo.toml
@@ -6,7 +6,7 @@ edition = "2018"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [[bench]]
-name = "benchmark"
+name = "userkind"
 harness = false
 
 [dependencies]
